Migrated the user profile script to TypeScript
